Rewrite in Python
While Bash is nice and simple for very simple tasks, it becomes
increasingly inadequate as features are added. In order to continue
improve the program, Python is now used.

This rewrite reimplements all the features previously supported and also
adds `output` selection. The user can now select an output (display).

Also, the edition is not prompted anymore. To edit a screenshot, the
user can simply use the default notification action, e.g. for Mako this
means clicking on the notification with its default configuration.